requirements.txt
appdirs==1.4.4
python-rtmidi==1.5.4
